# app/infra/http/swagger/docs/swagger.yaml
definitions:
  dto.CreateDownlinkRequest:
    properties:
      device_eui:
        type: string
      device_type:
        enum:
        - gateway
        - sensor
        type: string
      expires_at:
        type: string
      id:
        type: string
      payload:
        description: base64 encoded
        type: string
      type:
        enum:
        - config
        - command
        - firmware
        - ack
        type: string
    required:
    - device_eui
    - device_type
    - payload
    - type
    type: object
  dto.DownlinkResponse:
    properties:
      created_at:
        type: string
      device_eui:
        type: string
      device_type:
        type: string
      expires_at:
        type: string
      id:
        type: string
      payload:
        description: base64 encoded
        type: string
      retry_count:
        type: integer
      status:
        type: string
      type:
        type: string
    type: object
  dto.ErrorResponse:
    properties:
      error:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Downlink service for border gateways and sensors
  title: Bridgehead API
  version: "1.0"
paths:
  /ping:
    get:
      consumes:
      - application/json
      description: 'Ping me: and expect a response with Pong'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Use me to check if the service is healthy
      tags:
      - HealthCheck for the service
  /v1/downlinks:
    get:
      parameters:
      - description: device EUI
        in: query
        name: device_eui
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.DownlinkResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: List downlink requests by device EUI
      tags:
      - downlink
    post:
      consumes:
      - application/json
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateDownlinkRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.DownlinkResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Create downlink request
      tags:
      - downlink
  /v1/downlinks/{id}:
    delete:
      parameters:
      - description: request id
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Delete downlink request
      tags:
      - downlink
    get:
      parameters:
      - description: request id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DownlinkResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get downlink request
      tags:
      - downlink
swagger: "2.0"